Moved ADDI testing to new I instruction test file
